feat(analytics-docs): Refactor LiveLog components and improve modularity
