+++ packages/analytics-docs/src/components/LiveLog/constants.ts
@@ -1,0 +1,6 @@
+export const META_KEYS = ['version', 'commit', 'instanceId', 'sessionId', 'messageId'] as const;
+export const SHOW_META_IN_PAYLOAD_STORAGE_KEY = 'analytics-docs-live-log-show-meta-in-payload';
+
+export const NEW_EVENT_TIMEOUT = 20_000;
+export const LIVE_LOG_SIDEBAR_MIN_WIDTH = 280;
+export const LIVE_LOG_SIDEBAR_MAX_WIDTH = 800;
